backend/requirements.txt
requests==2.26.0
lxml==4.7.1
Pillow==8.4.0
aiohttp==3.8.1
gunicorn==20.1.0
aiohttp-swagger==1.0.16